fix type

+++ src/xtd_drawing/include/xtd/drawing/bitmap.hpp
@@ -19,7 +19,7 @@ namespace xtd {
       
       bitmap(int32_t width, int32_t height) : xtd::drawing::image(width, height) {}
       
-      drawing::color get_pixel(size_t x, size_t y) const;
+      drawing::color get_pixel(int32_t x, int32_t y) const;
     };
   }
 }
